+++ src/main/java/ru/job4j/dreamjob/service/SimpleVacancyService.java
@@ -2,6 +2,7 @@ package ru.job4j.dreamjob.service;
 
 import net.jcip.annotations.ThreadSafe;
 import org.springframework.stereotype.Service;
+import ru.job4j.dreamjob.dto.FileDto;
 import ru.job4j.dreamjob.model.Vacancy;
 import ru.job4j.dreamjob.repository.VacancyRepository;
 
@@ -13,24 +14,55 @@ import java.util.Optional;
 public class SimpleVacancyService implements VacancyService {
 
     private final VacancyRepository vacancyRepository;
+    private final FileService fileService;
 
-    public SimpleVacancyService(VacancyRepository vacancyRepository) {
+    public SimpleVacancyService(VacancyRepository vacancyRepository, FileService fileService) {
         this.vacancyRepository = vacancyRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public void save(Vacancy vacancy) {
+    public void save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
         vacancyRepository.save(vacancy);
     }
 
     @Override
     public boolean deleteById(int id) {
-        return vacancyRepository.deleteById(id);
+        var vacancyOptional = findById(id);
+        if (vacancyOptional.isEmpty()) {
+            return false;
+        }
+        var vacancy = vacancyOptional.get();
+        vacancyRepository.deleteById(id);
+
+        Optional.ofNullable(vacancy.getFileId())
+                .ifPresent(fileService::deleteById);
+
+        return true;
     }
 
     @Override
-    public boolean update(Vacancy vacancy) {
-        return vacancyRepository.update(vacancy);
+    public boolean update(Vacancy vacancy, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return vacancyRepository.update(vacancy);
+        }
+
+        Integer oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        var isUpdated = vacancyRepository.update(vacancy);
+
+        if (isUpdated && oldFileId != null) {
+            fileService.deleteById(oldFileId);
+        }
+
+        return isUpdated;
+    }
+
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        var file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
     @Override
